requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-appconf==1.0.5
django-cors-headers==3.13.0
django-imagekit==4.1.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
djoser==2.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
pandas==1.4.2
pilkit==2.0
Pillow==9.1.1
pycparser==2.21
PyJWT==2.4.0
pymongo==3.12.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
shortuuid==1.0.9
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.2.4
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
wincertstore==0.2
